backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
PyYAML>=6.0.1
requests>=2.31.0
websockets>=12.0
colorama>=0.4.6